Move loadFirebaseData() to db.js
